valen/class_9/app/requirements.txt
pandas==2.2.2
google-cloud-bigquery==3.25.0
pandas-gbq==0.23.1
openpyxl==3.1.5
